+++ src/openapi.rs
@@ -12,6 +12,7 @@
         crate::routes::http::shutdown,
         crate::routes::http::restart,
         crate::routes::http::echo,
+        crate::routes::http::get_current_machine,
         crate::routes::http::list_machines,
         crate::cron_jobs::list,
         crate::cron_jobs::create,
@@ -56,6 +57,7 @@
         crate::routes::http::RestartResponse,
         crate::routes::http::EchoRequest,
         crate::routes::http::EchoResponse,
+        crate::routes::http::MachineResponse,
         crate::global_lock::LockStatus,
         crate::routines::LockRequest,
     ))